+++ jkube-kit/resource/helm/src/main/java/org/eclipse/jkube/kit/resource/helm/HelmParameter.java
@@ -28,8 +28,15 @@ public class HelmParameter {
   public Parameter getParameter() {
     return parameter;
   }
-
+  
   public String getHelmName() {
     return parameter.getName().toLowerCase();
+  }
+
+  public String getHelmName(final boolean preserveCase) {
+    if(preserveCase) {
+      return parameter.getName();
+    }
+    return getHelmName();
   }
 }